+++ NEWS.md
@@ -2,6 +2,8 @@
 This is a patch (pre) release. It includes
 
 * Correcting some of the unit tests in `test-global-wrapper`.
+* Update the code in `nloptr.c` for compatibility with the `R` API, efficiency,
+and formatting.
 
 # nloptr 2.1.1